feat: add support for presentation tags
Check the presentation content for tags in the format:
link{title:"", pesenter:"<presenter_url>", viewer:"<viewer_url>"}
When such tags are detected in the presentatino content, a small button
is added to the presentation toolbar to play the url.
Title and viewer fields are mandatory, presenter is optional.
When there is no presenter field, the url specified in viewer field is
displayed to everyone.